Feature/pre 86 scale codec implemented (#54)
* pre-86 partial implementation

intermediate version, not for review yet. outcome::result finally applied.

* refactored

* tests added, bug fixed

* minor fixs

* minor fixes

+++ core/primitives/block_header.cpp
@@ -19,7 +19,7 @@ namespace kagome::primitives {
     return parent_hash_;
   }
 
-  size_t BlockHeader::number() const {
+  uint64_t BlockHeader::number() const {
     return number_;
   }
